backend/requirements.txt
altgraph==0.17.4
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
deep-translator==1.11.4
Django==5.2.4
django-tenants==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gTTS==2.5.4
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pefile==2023.2.7
psycopg2-binary==2.9.10
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyJWT==2.10.1
pypiwin32==223
PySocks==1.7.1
python-dotenv==1.1.1
pyttsx3==2.98
pywin32==310
pywin32-ctypes==0.2.3
requests==2.32.4
schedule==1.2.2
selenium==4.34.2
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2025.6.25
gunicorn
